Switch zcash and litecoin to use exchange rate based fee provider
