+++ bitcoincash/exchangerates.go
